// libs/features/violations/src/lib/shared/enums/document-transaction-id.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export enum DocumentTransactionId {
  MODIFY_VIOLATION_ID = '300347',
  CANCEL_VIOLATION_ID = '300348',
  REGISTER_CONTRIBUTOR_VIOLATION_ID = '300346',
  RAISE_VIOLATION_FO_ID = '300393',
  REGISTER_CONTRIBUTOR_VIOLATION_TYPE = 'REGISTER_CONTRIBUTOR_VIOLATION',
  REGISTER_CHANGE_ENGAGEMENT_VIOLATION = 'REGISTER_CHANGE_ENGAGEMENT_VIOLATION',
  MANUALLY_TRIGGERED_VIOLATION = 'MANUALLY_TRIGGERED_VIOLATION'
}