// libs/features/violations/src/lib/shared/enums/inspection-channel.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */
export enum InspectionChannel {
  E_INSPECTION = 'E-INSPECTION',
  RASED = 'RASED',
  SIMIS = 'SIMIS',
  HRSD = 'MINISTRY OF HUMAN RESOURCES AND SOCIAL DEVELOPMENT',
  KASHEF = 'KASHEF'
}